moving write_xga_config as a method of archive class
